// drr/src/UI/src/model/activityType.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * DRR API
 * OpenAPI spec version: 1.0.0
 */

export type ActivityType = typeof ActivityType[keyof typeof ActivityType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ActivityType = {
  Administration: 'Administration',
  ApprovalsPermitting: 'ApprovalsPermitting',
  Assessment: 'Assessment',
  Communications: 'Communications',
  Construction: 'Construction',
  ConstructionContractAward: 'ConstructionContractAward',
  ConstructionTender: 'ConstructionTender',
  Design: 'Design',
  FirstNationsEngagement: 'FirstNationsEngagement',
  LandAcquisition: 'LandAcquisition',
  Mapping: 'Mapping',
  AffectedPartiesEngagement: 'AffectedPartiesEngagement',
  PermitToConstruct: 'PermitToConstruct',
  Project: 'Project',
  ProjectPlanning: 'ProjectPlanning',
  CommunityEngagement: 'CommunityEngagement',
} as const;