// drr/src/UI/src/model/fundingStream.ts
/**
 * Generated by orval v7.1.1 🍺
 * Do not edit manually.
 * DRR API
 * OpenAPI spec version: 1.0.0
 */

export type FundingStream = typeof FundingStream[keyof typeof FundingStream];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const FundingStream = {
  Stream1: 'Stream1',
  Stream2: 'Stream2',
} as const;